requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
pillow==12.1.1
requests==2.32.5
urllib3==2.6.3
